use ccache for quick builds

+++ build-test-quick.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
-
-## Do not build vignettes. Allows quick tests 
+## Do not build vignettes. Allows quick tests
+## By default, we do not rm the .so, etc, and we can use ccache.
+## If you pass a second argument, we rm those, so we rebuild completely.
 V_R=$1
 
 V_ADA=$(cat ./OncoSimulR/DESCRIPTION | grep Version | cut -d' ' -f2)
@@ -32,15 +33,6 @@ rm ./OncoSimulR/R/Rplots.pdf
 rm ./OncoSimulR/vignettes/*~
 rm ./OncoSimulR/vignettes/Rplots.pdf
 rm ./OncoSimulR/vignettes/.Rhistory
-rm ./OncoSimulR/src/*.so
-rm ./OncoSimulR/src/*~
-rm ./OncoSimulR/src/*.o
-rm ./OncoSimulR/src/*.gcda
-rm ./OncoSimulR/src/*.gcno
-rm ./OncoSimulR/src/OncoSimulR.so
-rm ./OncoSimulR/src/OncoSimul.o
-rm ./OncoSimulR/src/OncoSimulR_init.o
-rm ./OncoSimulR/src/symbols.rds
 rm /OncoSimulR.Rcheck/* -r -f
 rm ./OncoSimulR.Rcheck/* -r -f
 # rm ./OncoSimulR/inst/doc/auto/*
@@ -58,13 +50,28 @@ rm ./OncoSimulR/vignettes/*.log
 rm ./OncoSimulR/vignettes/*.out
 rm ./OncoSimulR/vignettes/*.blg
 rm ./OncoSimulR/vignettes/*.synctex.*
+
+
+rm ./OncoSimulR/src/*~
+rm ./OncoSimulR/src/symbols.rds
+rm ./OncoSimulR/src/*.gcda
+rm ./OncoSimulR/src/*.gcno
+rm ./OncoSimulR/src/FitnessLandscape/*~
+if [ "$#" -ne 1 ]; then
+rm ./OncoSimulR/src/*.so
+rm ./OncoSimulR/src/*.o
+rm ./OncoSimulR/src/OncoSimulR.so
+rm ./OncoSimulR/src/OncoSimul.o
+rm ./OncoSimulR/src/OncoSimulR_init.o
 rm ./OncoSimulR/src/liblandscape.a
 rm ./OncoSimulR/src/fl_statistics 
 rm ./OncoSimulR/src/fl_generate
 rm ./OncoSimulR/src/fl_genchains
 rm ./OncoSimulR/src/FitnessLandscape/*.o
-rm ./OncoSimulR/src/FitnessLandscape/*~
 make -C ./OncoSimulR/src/FitnessLandscape clean
+fi
+
+
 
 ## We always do this, though it should not be necessary
 sed -i 's/^};$/}/' ./OncoSimulR/src/FitnessLandscape/generalized_chain.c
